# docs/swagger.yaml
openapi: 3.0.0
info:
  title: todoApp
  description: todoApp Sample for study
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local server
paths:
...
  /users/me:
    get:
      summary: ログインユーザーを取得
      description: ログインユーザーを取得します
      security:
        - cookieAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
      operationId: getMe
      tags:
        - users
    patch:
      summary: ログインユーザーを更新
      description: ログインユーザーを更新します
      security:
        - cookieAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad Request
        401:
          description: Unauthorized
      operationId: updateMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      tags:
        - users
  /users:
    post:
      summary: ユーザーを作成
      description: ユーザーを作成します
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad Request
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      tags:
        - users
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Loginに使用するCookie
  schemas:
...
    User:
      title: User
      type: object
      properties:
        id:
          type: string
          description: ユーザーUUID
          format: uuid
          x-go-type: uuid.UUID
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]{1,32}$'
          description: ユーザー名
      required:
        - id
        - name
        - createdAt
    UpdateUserRequest:
      title: UpdateUserRequest
      type: object
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]{1,32}$'
          description: ユーザー名
      required:
        - name
    CreateUserRequest:
      title: CreateUserRequest
      type: object
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]{1,32}$'
          description: ユーザー名
        password:
          type: string
          description: パスワード
      required:
        - name
        - password
...